Structured the AIAssistant components and added styling
