+++ src/components/Techstack.tsx
@@ -1,7 +1,7 @@
 import React, { useState } from "react";
 import { Box, Typography, Tabs, Tab, Paper, Rating, Tooltip, IconButton, useMediaQuery, useTheme } from "@mui/material";
 import { Code, Speed, ArrowBack, ArrowForward, PrecisionManufacturing } from "@mui/icons-material";
-import AIAssistant from "./AIAssistant";
+import AIAssistant from "./AIAssistantUI/AIAssistant";
 import useTabLayout from "../hooks/useTabLayout";
 import { 
   SiReact, SiTypescript, SiJavascript, SiNodedotjs, SiExpress,
