+++ src/types/assistant.ts
@@ -1,0 +1,4 @@
+export interface Message {
+  role: 'ai' | 'user';
+  text: string;
+}
